move ZKPGenerator/Validator; expect sync ZKP module & drop async/await; use functional options pattern for ZKPGenerator
