+++ ts_src/confidential.ts
@@ -1,18 +1,6 @@
 import * as crypto from './crypto';
 import { Output, ZERO } from './transaction';
-import type { ZKP } from '@vulpemventures/secp256k1-zkp';
-import {
-  Pset,
-  IssuanceBlindingArgs,
-  OutputBlindingArgs,
-  OwnedInput,
-  PsetBlindingGenerator,
-  PsetBlindingValidator,
-} from './psetv2';
-import { Slip77Interface } from 'slip77';
 import { ElementsValue } from './value';
-import { ECPairFactory, TinySecp256k1Interface } from 'ecpair';
-import * as randombytes from 'randombytes';
 
 export interface UnblindOutputResult {
   value: string;
@@ -28,28 +16,140 @@ export interface RangeProofInfoResult {
   maxValue: number;
 }
 
-export class Confidential {
-  constructor(private secp256k1Promise: Promise<ZKP>) {}
+type Ecdh = (pubkey: Buffer, scalar: Buffer) => Buffer;
 
-  async nonceHash(pubkey: Buffer, privkey: Buffer): Promise<Buffer> {
-    const { ecdh } = await this.secp256k1Promise;
-    return crypto.sha256(ecdh(pubkey, privkey));
+interface Ec {
+  prvkeyNegate: (key: Buffer) => Buffer;
+  prvkeyTweakAdd: (key: Buffer, tweak: Buffer) => Buffer;
+  prvkeyTweakMul: (key: Buffer, tweak: Buffer) => Buffer;
+}
+
+interface Generator {
+  generate: (seed: Buffer) => Buffer;
+  generateBlinded(key: Buffer, blind: Buffer): Buffer;
+  parse(input: Buffer): Buffer;
+  serialize(generator: Buffer): Buffer;
+}
+
+interface Pedersen {
+  commit(blindFactor: Buffer, value: string, generator: Buffer): Buffer;
+  commitSerialize(commitment: Buffer): Buffer;
+  commitParse(input: Buffer): Buffer;
+  blindSum(blinds: Array<Buffer>, nneg?: number): Buffer;
+  verifySum(commits: Array<Buffer>, negativeCommits: Array<Buffer>): boolean;
+  blindGeneratorBlindSum(
+    values: Array<string>,
+    nInputs: number,
+    blindGenerators: Array<Buffer>,
+    blindFactors: Array<Buffer>,
+  ): Buffer;
+}
+
+interface RangeProof {
+  info(proof: Buffer): {
+    exp: number;
+    mantissa: string;
+    minValue: string;
+    maxValue: string;
+  };
+  verify(
+    commit: Buffer,
+    proof: Buffer,
+    generator: Buffer,
+    extraCommit?: Buffer,
+  ): boolean;
+  sign(
+    commit: Buffer,
+    blind: Buffer,
+    nonce: Buffer,
+    value: string,
+    generator: Buffer,
+    minValue?: string,
+    base10Exp?: number,
+    minBits?: number,
+    message?: Buffer,
+    extraCommit?: Buffer,
+  ): Buffer;
+  rewind(
+    commit: Buffer,
+    proof: Buffer,
+    nonce: Buffer,
+    generator: Buffer,
+    extraCommit?: Buffer,
+  ): {
+    value: string;
+    minValue: string;
+    maxValue: string;
+    blindFactor: Buffer;
+    message: Buffer;
+  };
+}
+
+interface SurjectionProof {
+  serialize: (proof: {
+    nInputs: number;
+    usedInputs: Buffer;
+    data: Buffer;
+  }) => Buffer;
+  parse: (proof: Buffer) => {
+    nInputs: number;
+    usedInputs: Buffer;
+    data: Buffer;
+  };
+  initialize: (
+    inputTags: Array<Buffer>,
+    inputTagsToUse: number,
+    outputTag: Buffer,
+    maxIterations: number,
+    seed: Buffer,
+  ) => {
+    proof: { nInputs: number; usedInputs: Buffer; data: Buffer };
+    inputIndex: number;
+  };
+  generate: (
+    proof: { nInputs: number; usedInputs: Buffer; data: Buffer },
+    inputTags: Array<Buffer>,
+    outputTag: Buffer,
+    inputIndex: number,
+    inputBlindingKey: Buffer,
+    outputBlindingKey: Buffer,
+  ) => { nInputs: number; usedInputs: Buffer; data: Buffer };
+  verify: (
+    proof: { nInputs: number; usedInputs: Buffer; data: Buffer },
+    inputTags: Array<Buffer>,
+    outputTag: Buffer,
+  ) => boolean;
+}
+
+export interface ZKP {
+  ecdh: Ecdh;
+  ec: Ec;
+  surjectionproof: SurjectionProof;
+  rangeproof: RangeProof;
+  pedersen: Pedersen;
+  generator: Generator;
+}
+
+export class Confidential {
+  constructor(private zkp: ZKP) {}
+
+  nonceHash(pubkey: Buffer, privkey: Buffer): Buffer {
+    return crypto.sha256(this.zkp.ecdh(pubkey, privkey));
   }
 
-  async valueBlindingFactor(
+  valueBlindingFactor(
     inValues: string[],
     outValues: string[],
     inGenerators: Buffer[],
     outGenerators: Buffer[],
     inFactors: Buffer[],
     outFactors: Buffer[],
-  ): Promise<Buffer> {
-    const { pedersen } = await this.secp256k1Promise;
+  ): Buffer {
     const values = inValues.concat(outValues);
     const nInputs = inValues.length;
     const generators = inGenerators.concat(outGenerators);
     const factors = inFactors.concat(outFactors);
-    return pedersen.blindGeneratorBlindSum(
+    return this.zkp.pedersen.blindGeneratorBlindSum(
       values,
       nInputs,
       generators,
@@ -57,39 +157,32 @@ export class Confidential {
     );
   }
 
-  async valueCommitment(
-    value: string,
-    gen: Buffer,
-    factor: Buffer,
-  ): Promise<Buffer> {
-    const { generator, pedersen } = await this.secp256k1Promise;
+  valueCommitment(value: string, gen: Buffer, factor: Buffer): Buffer {
+    const { generator, pedersen } = this.zkp;
     const generatorParsed = generator.parse(gen);
     const commit = pedersen.commit(factor, value, generatorParsed);
     return pedersen.commitSerialize(commit);
   }
 
-  async assetCommitment(asset: Buffer, factor: Buffer): Promise<Buffer> {
-    const { generator } = await this.secp256k1Promise;
+  assetCommitment(asset: Buffer, factor: Buffer): Buffer {
+    const { generator } = this.zkp;
     const gen = generator.generateBlinded(asset, factor);
     return generator.serialize(gen);
   }
 
-  async unblindOutputWithKey(
+  unblindOutputWithKey(
     out: Output,
     blindingPrivKey: Buffer,
-  ): Promise<UnblindOutputResult> {
-    const nonce = await this.nonceHash(out.nonce, blindingPrivKey);
+  ): UnblindOutputResult {
+    const nonce = this.nonceHash(out.nonce, blindingPrivKey);
     return this.unblindOutputWithNonce(out, nonce);
   }
 
-  async unblindOutputWithNonce(
-    out: Output,
-    nonce: Buffer,
-  ): Promise<UnblindOutputResult> {
+  unblindOutputWithNonce(out: Output, nonce: Buffer): UnblindOutputResult {
     if (!out.rangeProof || out.rangeProof.length === 0) {
       throw new Error('Missing rangeproof to rewind');
     }
-    const secp = await this.secp256k1Promise;
+    const secp = this.zkp;
     const gen = secp.generator.parse(out.asset);
     const { value, blindFactor, message } = secp.rangeproof.rewind(
       out.value,
@@ -107,8 +200,8 @@ export class Confidential {
     };
   }
 
-  async rangeProofInfo(proof: Buffer): Promise<RangeProofInfoResult> {
-    const { rangeproof } = await this.secp256k1Promise;
+  rangeProofInfo(proof: Buffer): RangeProofInfoResult {
+    const { rangeproof } = this.zkp;
     const { exp, mantissa, minValue, maxValue } = rangeproof.info(proof);
     return {
       minValue: parseInt(minValue, 10),
@@ -121,7 +214,7 @@ export class Confidential {
   /**
    *  nonceHash from blinding key + ephemeral key and then rangeProof computation
    */
-  async rangeProofWithNonceHash(
+  rangeProofWithNonceHash(
     value: string,
     blindingPubkey: Buffer,
     ephemeralPrivkey: Buffer,
@@ -133,8 +226,8 @@ export class Confidential {
     minValue?: string,
     exp?: number,
     minBits?: number,
-  ): Promise<Buffer> {
-    const nonce = await this.nonceHash(blindingPubkey, ephemeralPrivkey);
+  ): Buffer {
+    const nonce = this.nonceHash(blindingPubkey, ephemeralPrivkey);
     return this.rangeProof(
       value,
       nonce,
@@ -149,13 +242,13 @@ export class Confidential {
     );
   }
 
-  async rangeProofVerify(
+  rangeProofVerify(
     valueCommit: Buffer,
     assetCommit: Buffer,
     proof: Buffer,
     script?: Buffer,
-  ): Promise<boolean> {
-    const { generator, pedersen, rangeproof } = await this.secp256k1Promise;
+  ): boolean {
+    const { generator, pedersen, rangeproof } = this.zkp;
     const gen = generator.parse(assetCommit);
     const commit = pedersen.commitParse(valueCommit);
     return rangeproof.verify(commit, proof, gen, script);
@@ -164,7 +257,7 @@ export class Confidential {
   /**
    *  rangeProof computation without nonceHash step.
    */
-  async rangeProof(
+  rangeProof(
     value: string,
     nonce: Buffer,
     asset: Buffer,
@@ -175,8 +268,8 @@ export class Confidential {
     minValue?: string,
     exp?: number,
     minBits?: number,
-  ): Promise<Buffer> {
-    const { generator, pedersen, rangeproof } = await this.secp256k1Promise;
+  ): Buffer {
+    const { generator, pedersen, rangeproof } = this.zkp;
 
     const gen = generator.generateBlinded(asset, assetBlindingFactor);
     const message = Buffer.concat([asset, assetBlindingFactor]);
@@ -200,14 +293,14 @@ export class Confidential {
     );
   }
 
-  async surjectionProof(
+  surjectionProof(
     outputAsset: Buffer,
     outputAssetBlindingFactor: Buffer,
     inputAssets: Buffer[],
     inputAssetBlindingFactors: Buffer[],
     seed: Buffer,
-  ): Promise<Buffer> {
-    const { generator, surjectionproof } = await this.secp256k1Promise;
+  ): Buffer {
+    const { generator, surjectionproof } = this.zkp;
     const outputGenerator = generator.generateBlinded(
       outputAsset,
       outputAssetBlindingFactor,
@@ -239,14 +332,14 @@ export class Confidential {
     return surjectionproof.serialize(proof);
   }
 
-  async surjectionProofVerify(
+  surjectionProofVerify(
     inAssets: Buffer[],
     inAssetBlinders: Buffer[],
     outAsset: Buffer,
     outAssetBlinder: Buffer,
     proof: Buffer,
-  ): Promise<boolean> {
-    const { generator, surjectionproof } = await this.secp256k1Promise;
+  ): boolean {
+    const { generator, surjectionproof } = this.zkp;
     const inGenerators = inAssets.map((v, i) =>
       generator.generateBlinded(v, inAssetBlinders[i]),
     );
@@ -255,28 +348,27 @@ export class Confidential {
     return surjectionproof.verify(sProof, inGenerators, outGenerator);
   }
 
-  async blindValueProof(
+  blindValueProof(
     value: string,
     valueCommit: Buffer,
     assetCommit: Buffer,
     valueBlinder: Buffer,
-    opts?: RngOpts,
-  ): Promise<Buffer> {
-    const { generator, pedersen, rangeproof } = await this.secp256k1Promise;
+    nonce: Buffer,
+  ): Buffer {
+    const { generator, pedersen, rangeproof } = this.zkp;
 
     const gen = generator.parse(assetCommit);
     const commit = pedersen.commitParse(valueCommit);
-    const nonce = randomBytes(opts);
 
     return rangeproof.sign(commit, valueBlinder, nonce, value, gen, value, -1);
   }
 
-  async blindAssetProof(
+  blindAssetProof(
     asset: Buffer,
     assetCommit: Buffer,
     assetBlinder: Buffer,
-  ): Promise<Buffer> {
-    const { generator, surjectionproof } = await this.secp256k1Promise;
+  ): Buffer {
+    const { generator, surjectionproof } = this.zkp;
 
     const nInputsToUse = 1;
     const maxIterations = 100;
@@ -304,688 +396,17 @@ export class Confidential {
     return surjectionproof.serialize(proof);
   }
 
-  async assetBlindProofVerify(
+  assetBlindProofVerify(
     asset: Buffer,
     assetCommit: Buffer,
     proof: Buffer,
-  ): Promise<boolean> {
-    const { generator, surjectionproof } = await this.secp256k1Promise;
+  ): boolean {
+    const { generator, surjectionproof } = this.zkp;
     const inGenerators = [generator.generate(asset)];
     const outGenerator = generator.parse(assetCommit);
     const sProof = surjectionproof.parse(proof);
     return surjectionproof.verify(sProof, inGenerators, outGenerator);
   }
-}
-
-// const secp256k1Promise = secp256k1();
-
-interface RngOpts {
-  rng?(arg0: number): Buffer;
-}
-
-export type KeysGenerator = (opts?: RngOpts) => {
-  publicKey: Buffer;
-  privateKey: Buffer;
-};
-
-export class ZKPValidator implements PsetBlindingValidator {
-  constructor(private zkpLib: Promise<any>) {}
-
-  async verifyValueRangeProof(
-    valueCommit: Buffer,
-    assetCommit: Buffer,
-    proof: Buffer,
-    script: Buffer,
-  ): Promise<boolean> {
-    try {
-      const confidential = new Confidential(this.zkpLib);
-      return await confidential.rangeProofVerify(
-        valueCommit,
-        assetCommit,
-        proof,
-        script,
-      );
-    } catch (ignore) {
-      return false;
-    }
-  }
-
-  async verifyAssetSurjectionProof(
-    inAssets: Buffer[],
-    inAssetBlinders: Buffer[],
-    outAsset: Buffer,
-    outAssetBlinder: Buffer,
-    proof: Buffer,
-  ): Promise<boolean> {
-    try {
-      const confidential = new Confidential(this.zkpLib);
-      return await confidential.surjectionProofVerify(
-        inAssets,
-        inAssetBlinders,
-        outAsset,
-        outAssetBlinder,
-        proof,
-      );
-    } catch (ignore) {
-      return false;
-    }
-  }
-
-  async verifyBlindValueProof(
-    valueCommit: Buffer,
-    assetCommit: Buffer,
-    proof: Buffer,
-  ): Promise<boolean> {
-    try {
-      const confidential = new Confidential(this.zkpLib);
-      return await confidential.rangeProofVerify(
-        valueCommit,
-        assetCommit,
-        proof,
-      );
-    } catch (ignore) {
-      return false;
-    }
-  }
-
-  async verifyBlindAssetProof(
-    asset: Buffer,
-    assetCommit: Buffer,
-    proof: Buffer,
-  ): Promise<boolean> {
-    try {
-      const confidential = new Confidential(this.zkpLib);
-      return await confidential.assetBlindProofVerify(
-        asset,
-        assetCommit,
-        proof,
-      );
-    } catch (ignore) {
-      return false;
-    }
-  }
-}
-
-export class ZKPGenerator implements PsetBlindingGenerator {
-  static fromOwnedInputs(
-    ownedInputs: OwnedInput[],
-    zkpLib: Promise<any>,
-  ): ZKPGenerator {
-    const bg = new ZKPGenerator(zkpLib);
-    bg.ownedInputs = ownedInputs;
-    return bg;
-  }
-
-  static fromInBlindingKeys(
-    inBlindingKeys: Buffer[],
-    zkpLib: Promise<any>,
-  ): ZKPGenerator {
-    const bg = new ZKPGenerator(zkpLib);
-    bg.inBlindingKeys = inBlindingKeys;
-    return bg;
-  }
-
-  static fromMasterBlindingKey(
-    masterKey: Slip77Interface,
-    zkpLib: Promise<any>,
-  ): ZKPGenerator {
-    const bg = new ZKPGenerator(zkpLib);
-    bg.masterBlindingKey = masterKey;
-    return bg;
-  }
-
-  static ECCKeysGenerator(ec: TinySecp256k1Interface): KeysGenerator {
-    return (opts?: RngOpts) => {
-      const privateKey = randomBytes(opts);
-      const publicKey = ECPairFactory(ec).fromPrivateKey(privateKey).publicKey;
-      return {
-        privateKey,
-        publicKey,
-      };
-    };
-  }
-
-  ownedInputs?: OwnedInput[];
-  inBlindingKeys?: Buffer[];
-  masterBlindingKey?: Slip77Interface;
-  opts?: RngOpts;
-
-  private constructor(private secp256k1Promise: Promise<any>) {}
-
-  async computeAndAddToScalarOffset(
-    scalar: Buffer,
-    value: string,
-    assetBlinder: Buffer,
-    valueBlinder: Buffer,
-  ): Promise<Buffer> {
-    // If both asset and value blinders are null, 0 is added to the offset, so nothing actually happens
-    if (assetBlinder.equals(ZERO) && valueBlinder.equals(ZERO)) {
-      return scalar.slice();
-    }
-
-    const scalarOffset = await this.calculateScalarOffset(
-      value,
-      assetBlinder,
-      valueBlinder,
-    );
-
-    // When we start out, the result (a) is 0, so just set it to the scalar we just computed.
-    if (scalar.equals(ZERO)) {
-      return scalarOffset;
-    }
-
-    const { ec } = await this.secp256k1Promise;
-    const negScalarOffset = ec.prvkeyNegate(scalarOffset);
-
-    if (scalar.equals(negScalarOffset)) {
-      return ZERO;
-    }
-
-    return ec.prvkeyTweakAdd(scalar, scalarOffset);
-  }
-
-  async subtractScalars(
-    inputScalar: Buffer,
-    outputScalar: Buffer,
-  ): Promise<Buffer> {
-    if (outputScalar.equals(ZERO)) {
-      return inputScalar.slice();
-    }
-    const { ec } = await this.secp256k1Promise;
-    const negOutputScalar = ec.prvkeyNegate(outputScalar);
-    if (inputScalar.equals(ZERO)) {
-      return negOutputScalar;
-    }
-    return ec.prvkeyTweakAdd(inputScalar, negOutputScalar);
-  }
-
-  async lastValueCommitment(
-    value: string,
-    asset: Buffer,
-    blinder: Buffer,
-  ): Promise<Buffer> {
-    const confidential = new Confidential(this.secp256k1Promise);
-    return confidential.valueCommitment(value, asset, blinder);
-  }
-
-  async lastBlindValueProof(
-    value: string,
-    valueCommit: Buffer,
-    assetCommit: Buffer,
-    blinder: Buffer,
-  ): Promise<Buffer> {
-    const confidential = new Confidential(this.secp256k1Promise);
-    return confidential.blindValueProof(
-      value,
-      valueCommit,
-      assetCommit,
-      blinder,
-    );
-  }
-
-  async lastValueRangeProof(
-    value: string,
-    asset: Buffer,
-    valueCommit: Buffer,
-    valueBlinder: Buffer,
-    assetBlinder: Buffer,
-    script: Buffer,
-    nonce: Buffer,
-  ): Promise<Buffer> {
-    const confidential = new Confidential(this.secp256k1Promise);
-    return confidential.rangeProof(
-      value,
-      nonce,
-      asset,
-      assetBlinder,
-      valueBlinder,
-      valueCommit,
-      script,
-    );
-  }
-
-  async unblindInputs(pset: Pset, inIndexes?: number[]): Promise<OwnedInput[]> {
-    validatePset(pset);
-    if (inIndexes!) {
-      validateInIndexes(pset, inIndexes);
-    }
-
-    const inputIndexes =
-      inIndexes || Array.from({ length: pset.globals.inputCount }, (_, i) => i);
-
-    if (this.ownedInputs! && this.ownedInputs!.length > 0) {
-      return this.ownedInputs!;
-    }
-
-    const revealedInputs = await Promise.all(
-      inputIndexes.map(async (i) => {
-        const prevout = pset.inputs[i].getUtxo();
-        const revealedInput = await this.unblindUtxo(prevout!);
-        revealedInput.index = i;
-        return revealedInput;
-      }),
-    );
-    this.ownedInputs = revealedInputs;
-    return revealedInputs;
-  }
-
-  async blindIssuances(
-    pset: Pset,
-    blindingKeysByIndex: Record<number, Buffer>,
-  ): Promise<IssuanceBlindingArgs[]> {
-    validatePset(pset);
-    validateBlindingKeysByIndex(pset, blindingKeysByIndex);
-
-    const confidential = new Confidential(this.secp256k1Promise);
-
-    return Promise.all(
-      Object.entries(blindingKeysByIndex).map(async ([i, key]) => {
-        const input = pset.inputs[parseInt(i, 10)];
-
-        let blindingArgs = {} as IssuanceBlindingArgs;
-        if (input.issuanceValue! > 0) {
-          const value = input.issuanceValue!.toString(10);
-          const asset = input.getIssuanceAssetHash()!;
-          const blinder = randomBytes(this.opts);
-
-          const assetCommit = await confidential.assetCommitment(asset, ZERO);
-          const valueCommit = await confidential.valueCommitment(
-            value,
-            assetCommit,
-            blinder,
-          );
-          const blindproof = await confidential.blindValueProof(
-            value,
-            valueCommit,
-            assetCommit,
-            blinder,
-          );
-          const rangeproof = await confidential.rangeProof(
-            value,
-            key,
-            asset,
-            ZERO,
-            blinder,
-            valueCommit,
-            Buffer.from([]),
-          );
-
-          blindingArgs = {
-            ...blindingArgs,
-            index: parseInt(i, 10),
-            issuanceAsset: asset,
-            issuanceValueCommitment: valueCommit,
-            issuanceValueRangeProof: rangeproof,
-            issuanceValueBlindProof: blindproof,
-            issuanceValueBlinder: blinder,
-          };
-        }
-
-        if (input.issuanceInflationKeys! > 0) {
-          const token = input.issuanceInflationKeys!.toString(10);
-          const asset = input.getIssuanceInflationKeysHash(true)!;
-          const blinder = randomBytes(this.opts);
-          const assetCommit = await confidential.assetCommitment(asset, ZERO);
-          const tokenCommit = await confidential.valueCommitment(
-            token,
-            assetCommit,
-            blinder,
-          );
-          const blindproof = await confidential.blindValueProof(
-            token,
-            tokenCommit,
-            assetCommit,
-            blinder,
-          );
-          const rangeproof = await confidential.rangeProof(
-            token,
-            key,
-            asset,
-            ZERO,
-            blinder,
-            tokenCommit,
-            Buffer.from([]),
-          );
-
-          blindingArgs = {
-            ...blindingArgs,
-            issuanceToken: asset,
-            issuanceTokenCommitment: tokenCommit,
-            issuanceTokenRangeProof: rangeproof,
-            issuanceTokenBlindProof: blindproof,
-            issuanceTokenBlinder: blinder,
-          };
-        }
-
-        return blindingArgs;
-      }),
-    );
-  }
-
-  async blindOutputs(
-    pset: Pset,
-    keysGenerator: KeysGenerator,
-    outIndexes?: number[],
-    blindedIssuances?: IssuanceBlindingArgs[],
-  ): Promise<OutputBlindingArgs[]> {
-    validatePset(pset);
-    if (outIndexes!) {
-      validateOutIndexes(pset, outIndexes);
-    }
-    if (blindedIssuances! && blindedIssuances!.length > 0) {
-      validateBlindedIssuances(pset, blindedIssuances);
-    }
-
-    const outputIndexes =
-      outIndexes! && outIndexes.length > 0
-        ? outIndexes
-        : pset.outputs.reduce(
-            (arr: number[], out, i) => (
-              out.needsBlinding() && arr.push(i), arr
-            ),
-            [],
-          );
-
-    const { assets, assetBlinders } = await this.getInputAssetsAndBlinders(
-      pset,
-      blindedIssuances,
-    );
-
-    const confidential = new Confidential(this.secp256k1Promise);
-    return Promise.all(
-      outputIndexes.map(async (i) => {
-        const output = pset.outputs[i];
-        const assetBlinder = randomBytes(this.opts);
-        const valueBlinder = randomBytes(this.opts);
-        const seed = randomBytes(this.opts);
-        const value = output.value!.toString(10);
-        const assetCommit = await confidential.assetCommitment(
-          output.asset!,
-          assetBlinder,
-        );
-        const valueCommit = await confidential.valueCommitment(
-          value,
-          assetCommit,
-          valueBlinder,
-        );
-        const ephemeralKeyPair = keysGenerator();
-        const nonceCommitment = ephemeralKeyPair.publicKey;
-        const ecdhNonce = await confidential.nonceHash(
-          output.blindingPubkey!,
-          ephemeralKeyPair.privateKey,
-        );
-        const script = output.script || Buffer.from([]);
-        const rangeproof = await confidential.rangeProof(
-          value,
-          ecdhNonce,
-          output.asset!,
-          assetBlinder,
-          valueBlinder,
-          valueCommit,
-          script,
-        );
-        const surjectionproof = await confidential.surjectionProof(
-          output.asset!,
-          assetBlinder,
-          assets,
-          assetBlinders,
-          seed,
-        );
-        const valueBlindProof = await confidential.blindValueProof(
-          value,
-          valueCommit,
-          assetCommit,
-          valueBlinder,
-        );
-        const assetBlindProof = await confidential.blindAssetProof(
-          output.asset!,
-          assetCommit,
-          assetBlinder,
-        );
-
-        return {
-          index: i,
-          nonce: ecdhNonce,
-          nonceCommitment,
-          valueCommitment: valueCommit,
-          assetCommitment: assetCommit,
-          valueRangeProof: rangeproof,
-          assetSurjectionProof: surjectionproof,
-          valueBlindProof,
-          assetBlindProof,
-          valueBlinder,
-          assetBlinder,
-        };
-      }),
-    );
-  }
-
-  private async calculateScalarOffset(
-    value: string,
-    assetBlinder: Buffer,
-    valueBlinder: Buffer,
-  ): Promise<Buffer> {
-    if (assetBlinder.equals(ZERO)) {
-      return valueBlinder.slice();
-    }
-    if (value === '0') {
-      return valueBlinder.slice();
-    }
-
-    const { ec } = await this.secp256k1Promise;
-    const val = Buffer.alloc(32, 0);
-    val.writeBigUInt64BE(BigInt(value), 24);
-    const result = ec.prvkeyTweakMul(assetBlinder, val);
-    if (valueBlinder.length === 0) {
-      throw new Error('Missing value blinder');
-    }
-
-    const negVb = ec.prvkeyNegate(valueBlinder);
-
-    if (negVb.equals(result)) {
-      return ZERO;
-    }
-
-    return ec.prvkeyTweakAdd(result, valueBlinder);
-  }
-
-  private async unblindUtxo(out: Output): Promise<OwnedInput> {
-    if (out.nonce.length === 1) {
-      return {
-        index: 0,
-        value: ElementsValue.fromBytes(out.value).number.toString(10),
-        asset: out.asset.slice(1),
-        valueBlindingFactor: ZERO,
-        assetBlindingFactor: ZERO,
-      };
-    }
-
-    if (!this.inBlindingKeys && !this.masterBlindingKey) {
-      throw new Error(
-        'Missing either input private blinding keys or SLIP-77 master blinding key',
-      );
-    }
-
-    const keys = this.inBlindingKeys
-      ? this.inBlindingKeys
-      : [this.masterBlindingKey!.derive(out.script).privateKey!];
-
-    for (const key of keys) {
-      try {
-        const confidential = new Confidential(this.secp256k1Promise);
-        const revealed = await confidential.unblindOutputWithKey(out, key);
-        return {
-          index: 0,
-          value: revealed.value,
-          asset: revealed.asset,
-          valueBlindingFactor: revealed.valueBlindingFactor,
-          assetBlindingFactor: revealed.assetBlindingFactor,
-        };
-      } catch (ignore) {}
-    }
-
-    throw new Error('Could not unblind output with any blinding key');
-  }
-
-  private async getInputAssetsAndBlinders(
-    pset: Pset,
-    issuanceBlindingArgs?: IssuanceBlindingArgs[],
-  ): Promise<{
-    assets: Buffer[];
-    assetBlinders: Buffer[];
-  }> {
-    const unblindedIns = await this.maybeUnblindInUtxos(pset);
-    pset.inputs.forEach((input, i) => {
-      if (input.hasIssuance()) {
-        unblindedIns.push({
-          value: '',
-          valueBlindingFactor: Buffer.from([]),
-          asset: input.getIssuanceAssetHash()!,
-          assetBlindingFactor: ZERO,
-        });
-        if (input.issuanceInflationKeys! > 0) {
-          const isBlindedIssuance =
-            issuanceBlindingArgs! &&
-            issuanceBlindingArgs.find(({ index }) => index === i) !== undefined;
-          unblindedIns.push({
-            value: '',
-            valueBlindingFactor: Buffer.from([]),
-            asset: input.getIssuanceInflationKeysHash(isBlindedIssuance)!,
-            assetBlindingFactor: ZERO,
-          });
-        }
-      }
-    });
-
-    const assets = [] as Buffer[];
-    const assetBlinders = [] as Buffer[];
-
-    unblindedIns.forEach(({ asset, assetBlindingFactor }) => {
-      assets.push(asset);
-      assetBlinders.push(assetBlindingFactor);
-    });
-
-    return { assets, assetBlinders };
-  }
-
-  private async maybeUnblindInUtxos(
-    pset: Pset,
-  ): Promise<UnblindOutputResult[]> {
-    if (this.ownedInputs! && this.ownedInputs!.length > 0) {
-      return pset.inputs.map((input, i) => {
-        const ownedInput = this.ownedInputs!.find(({ index }) => index === i);
-        if (ownedInput!) {
-          return {
-            value: '',
-            valueBlindingFactor: Buffer.from([]),
-            asset: ownedInput!.asset,
-            assetBlindingFactor: ownedInput!.assetBlindingFactor,
-          };
-        }
-        return {
-          value: '',
-          valueBlindingFactor: Buffer.from([]),
-          asset: input.getUtxo()!.asset,
-          assetBlindingFactor: ZERO,
-        };
-      });
-    }
-
-    if (!this.inBlindingKeys && !this.masterBlindingKey) {
-      throw new Error(
-        'Missing either input private blinding keys or SLIP-77 master blinding key',
-      );
-    }
-
-    return Promise.all(
-      pset.inputs.map(async (input) => {
-        const prevout = input.getUtxo()!;
-        try {
-          const revealed = await this.unblindUtxo(prevout);
-          return {
-            value: revealed.value,
-            asset: revealed.asset,
-            valueBlindingFactor: revealed.valueBlindingFactor,
-            assetBlindingFactor: revealed.assetBlindingFactor,
-          };
-        } catch (ignore) {
-          return {
-            value: '',
-            asset: prevout.asset,
-            valueBlindingFactor: Buffer.from([]),
-            assetBlindingFactor: ZERO,
-          };
-        }
-      }),
-    );
-  }
-}
-
-function validatePset(pset: Pset): void {
-  pset.sanityCheck();
-
-  pset.inputs.forEach((input, i) => {
-    if (!input.getUtxo()) {
-      throw new Error('Missing (non-)witness utxo for input ' + i);
-    }
-  });
-}
-
-function validateInIndexes(pset: Pset, inIndexes: number[]): void {
-  if (inIndexes.length > 0) {
-    inIndexes.forEach((i) => {
-      if (i < 0 || i >= pset.globals.inputCount) {
-        throw new Error('Input index out of range');
-      }
-    });
-  }
-}
-
-function validateOutIndexes(pset: Pset, outIndexes: number[]): void {
-  if (outIndexes.length > 0) {
-    outIndexes.forEach((i) => {
-      if (i < 0 || i >= pset.globals.outputCount) {
-        throw new Error('Output index out of range');
-      }
-    });
-  }
-}
-
-function validateBlindingKeysByIndex(
-  pset: Pset,
-  keys: Record<number, Buffer>,
-): void {
-  Object.entries(keys).forEach(([k, v]) => {
-    const i = parseInt(k, 10);
-    if (i < 0 || i >= pset.globals.inputCount) {
-      throw new Error('Input index out of range');
-    }
-    if (!pset.inputs[i].hasIssuance()) {
-      throw new Error('Input does not have any issuance to blind');
-    }
-    if (v.length !== 32) {
-      throw new Error('Invalid private blinding key length for input ' + i);
-    }
-  });
-}
-
-function validateBlindedIssuances(
-  pset: Pset,
-  blindedIssuances: IssuanceBlindingArgs[],
-): void {
-  if (blindedIssuances.length > 0) {
-    blindedIssuances.forEach((issuance) => {
-      if (issuance.index < 0 || issuance.index >= pset.globals.inputCount) {
-        throw new Error('Input index of blinded issuance is out of range');
-      }
-    });
-  }
-}
-
-function randomBytes(options?: RngOpts): Buffer {
-  if (options === undefined) options = {};
-  const rng = options.rng || randombytes.default;
-  return rng(32);
 }
 
 export function confidentialValueToSatoshi(value: Buffer): number {
