+++ ts_src/index.ts
@@ -5,7 +5,6 @@ import * as payments from './payments';
 import * as script from './script';
 import * as issuance from './issuance';
 import * as bip341 from './bip341';
-
 export { address, crypto, networks, payments, script, issuance, bip341 };
 export { TaggedHashPrefix } from './crypto';
 export { OPS as opcodes } from './ops';
@@ -14,7 +13,6 @@ export {
   Output as TxOutput,
   Transaction,
 } from './transaction';
-export { Network as NetworkExtended } from './networks';
 export * from './asset';
 export * from './value';
 export * from './psetv2';
