+++ ts_src/psetv2/blinder.ts
@@ -1,6 +1,7 @@
 import { UnblindOutputResult } from '../confidential';
 import { ZERO } from '../transaction';
 import { Pset } from './pset';
+import { ZKPGenerator, ZKPValidator } from './zkp';
 
 export interface IssuanceBlindingArgs {
   index: number;
@@ -32,73 +33,17 @@ export interface OutputBlindingArgs {
 
 export type OwnedInput = { index: number } & UnblindOutputResult;
 
-export interface PsetBlindingGenerator {
-  computeAndAddToScalarOffset(
-    scalar: Buffer,
-    value: string,
-    assetBlinder: Buffer,
-    valueBlinder: Buffer,
-  ): Promise<Buffer>;
-  subtractScalars(inputScalar: Buffer, outputScalar: Buffer): Promise<Buffer>;
-  lastValueCommitment(
-    value: string,
-    asset: Buffer,
-    blinder: Buffer,
-  ): Promise<Buffer>;
-  lastBlindValueProof(
-    value: string,
-    valueCommitment: Buffer,
-    assetCommitment: Buffer,
-    blinder: Buffer,
-  ): Promise<Buffer>;
-  lastValueRangeProof(
-    value: string,
-    asset: Buffer,
-    valueCommitment: Buffer,
-    valueBlinder: Buffer,
-    assetBlinder: Buffer,
-    script: Buffer,
-    nonce: Buffer,
-  ): Promise<Buffer>;
-}
-
-export interface PsetBlindingValidator {
-  verifyValueRangeProof(
-    valueCommitment: Buffer,
-    assetCommitment: Buffer,
-    proof: Buffer,
-    script: Buffer,
-  ): Promise<boolean>;
-  verifyAssetSurjectionProof(
-    inAssets: Buffer[],
-    inAssetBlinders: Buffer[],
-    outAsset: Buffer,
-    outAssetBlinder: Buffer,
-    proof: Buffer,
-  ): Promise<boolean>;
-  verifyBlindValueProof(
-    valueCommitment: Buffer,
-    assetCommitment: Buffer,
-    proof: Buffer,
-  ): Promise<boolean>;
-  verifyBlindAssetProof(
-    asset: Buffer,
-    assetCommitment: Buffer,
-    proof: Buffer,
-  ): Promise<boolean>;
-}
-
 export class Blinder {
   pset: Pset;
   ownedInputs: OwnedInput[];
-  blindingValidator: PsetBlindingValidator;
-  blindingGenerator: PsetBlindingGenerator;
+  blindingValidator: ZKPValidator;
+  blindingGenerator: ZKPGenerator;
 
   constructor(
     pset: Pset,
     ownedInputs: OwnedInput[],
-    validator: PsetBlindingValidator,
-    generator: PsetBlindingGenerator,
+    validator: ZKPValidator,
+    generator: ZKPGenerator,
   ) {
     if (ownedInputs.length === 0) {
       throw new Error('Missing owned inputs');
